fixing recreate_parent_column
The function would change the order of appearance of the parent columns and flag them as changes.
In some surveys, options are the same but the order is different. The function would changed it  and later flags changes while it is this function that changed the order.
I don't know why the order is different than the order of the columns,  possible options: the choices order changed, the fill in order

Also correcting the initial for loop condition as it was repeating for each options and not each select multiple type.